+++ src/test/java/step2/domain/NameTest.java
@@ -1,4 +1,4 @@
-package step2;
+package step2.domain;
 
 import org.junit.jupiter.api.Test;
